Scorporated the conversion functions from ett.py and tte.py and the conversion coordinates from constants.py into emotion_conversion.py. Now constants.py it's interface_tags.py, since it only keeps tags information as interface to other groups.
