+++ SmartApp.ENLP/interface_tags.py
@@ -1,15 +1,3 @@
-# emotion : (valence, arousal)
-EMOTIONS_COORD = {
-    'Excited' : (0.5,0.8),
-    'Angry' : (-0.8,0.5),
-    'Bored' : (-0.5,-0.8),
-    'Happy' : (0.8,0.5),
-    'Fear' : (-0.6,0.8),
-    'Sarcasm' : (0.,0.3),
-    'Sad' : (-0.9,-0.6),
-    'Neutral' : (0.,0.)
-}
-
 #READ FROM KB
 TAG_USER_TRANSCRIPT = "text_f_audio" # user query transcript
 TAG_ANSWER = "NLP_Answer" # nlp answer to provide BEFORE addition of emotions
